+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## vNEXT
+- Remove `smock` from unit tests:
+    - IexecPocoBoost (#148)
 - Migrate unit test files to Typescript & Hardhat:
     - ERC1154 (#145, #146, #147)
     - IexecEscrowToken (#141, #143)